+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.7.29
 
+* added `wait_until_equal` to `AtomicNumber`